requirements.txt
# Requirements for Meta Oracle System

# Core dependencies
pandas>=2.0.0
pyarrow>=12.0.0

# Environment variable management
python-dotenv>=1.0.0

# Configuration management
pyyaml>=6.0.0

# LLM providers
anthropic>=0.18.0
openai>=1.0.0
google-genai>=1.0.0

# Optional: for enhanced data analysis
numpy>=1.24.0
matplotlib>=3.7.0

# Optional: for notebook exploration
jupyter>=1.0.0
ipykernel>=6.0.0
